+++ __tests__/Car.test.js
@@ -1,5 +1,5 @@
 import Car from "../src/domain/Car.js";
-import { ERROR_MESSAGE } from "../src/config/constants.js";
+import { ERROR_MESSAGE } from "../src/config/errorConstants.js";
 
 describe("Car 객체 테스트", () => {
   test.each(["happy", "h"])("자동차 생성 테스트", (name) => {
