+++ src/config/errorConstants.js
@@ -1,0 +1,11 @@
+export const ERROR_MESSAGE = {
+  CAR_NAME_MAX_LENGTH: "[ERROR] 자동차 이름은 5자 이하여야 합니다.",
+  CAR_NAME_MIN_LENGTH: "[ERROR] 자동차 이름은 최소 1자 이상이여야 합니다.",
+  CAR_NAME_KOREAN_AND_ENGLISH:
+    "[ERROR] 자동차 이름은 영어와 숫자만 가능합니다.",
+  CAR_MAX_COUNT: "[ERROR] 경주 가능한 최대 자동차 개수는 10개입니다.",
+  CAR_NAME_DUPLICATE: "[ERROR] 자동차 이름이 중복됩니다.",
+  INVALID_ATTEMPT_NUMBER: "[ERROR] 시도 횟수는 숫자여야 합니다.",
+  ATTEMPT_TOO_LOW: "[ERROR] 시도 횟수는 1 이상의 정수여야 합니다.",
+  ATTEMPT_TOO_HIGH: "[ERROR] 시도 횟수는 최대 50회까지 가능합니다.",
+};
