+++ src/config/ioConstans.js
@@ -1,0 +1,10 @@
+export const INPUT_MESSAGE = {
+  CAR_NAME_INPUT:
+    "경주할 자동차 이름을 입력하세요(이름은 쉼표(,)를 기준으로 구분).\n",
+  ATTEMPT_INPUT: "시도할 횟수는 몇 회인가요?\n",
+};
+
+export const OUTPUT_MESSAGE = {
+  RACE_RESULT_TITLE: "\n실행 결과",
+  WINNER_TITLE: "최종 우승자: ",
+};
